Reimplement HashPath core functionality as a library
This takes the guts of the heyday/silverstripe-hashpath package and
refactors it into something that is reuseable without using SilverStripe.
Previously I've created one-off modified versions of the SilverStripe
HashPath implementation for use in Symfony and Phalcon PHP projects,
which was a never a great longterm solution.

This implementation of HashPath leaves the responsibility of finding
filesystem and web paths up to user code, since these vary depending
on framework. It's tempting to add a utility/wrapper to the library
that takes a document root path and hashes files relative to that;
handling the conversion between filesystem and web paths automatically,
since that would likely cover the main use case.