Make PathModifier::joinPaths public
The reference framework integration of this library (heyday/silverstripe-hashpath)
requires joining a document root and web path together to make an absolute
filesystem path. It is almost guaranteed that user code will need to join
paths to provide to an instance of Hasher, so it seems sensible to expose
joinPaths given that PHP has no built-in solution for joining paths safely.